backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
sqlalchemy==2.0.35
alembic==1.13.3
pydantic==2.9.2
pydantic-settings==2.4.0
python-multipart==0.0.9
passlib[bcrypt]==1.7.4
pyjwt==2.9.0
python-slugify==8.0.4
email-validator==2.2.0
